Split scalar module into submodules
