+++ probdiffeq/implementations/blockdiag/corr.py
@@ -5,7 +5,8 @@ import jax
 import jax.numpy as jnp
 
 from probdiffeq.implementations import _collections, cubature
-from probdiffeq.implementations.scalar import _scalar
+from probdiffeq.implementations.scalar import _vars as scalar_vars
+from probdiffeq.implementations.scalar import corr as scalar_corr
 
 _SLR1CacheType = Tuple[Callable]
 """Type-variable for the correction-cache."""
@@ -31,7 +32,7 @@ class BlockDiagStatisticalFirstOrder(_collections.AbstractCorrection):
         super().__init__(ode_order=ode_order)
         self.ode_shape = ode_shape
 
-        self._mm = _scalar.StatisticalFirstOrder(
+        self._mm = scalar_corr.StatisticalFirstOrder(
             ode_order=ode_order, cubature_rule=cubature_rule
         )
 
@@ -67,15 +68,17 @@ class BlockDiagStatisticalFirstOrder(_collections.AbstractCorrection):
         cache = (vmap_f,)
 
         # Evaluate vector field at sigma-points
-        sigma_points_fn = jax.vmap(_scalar.StatisticalFirstOrder.transform_sigma_points)
+        sigma_points_fn = jax.vmap(
+            scalar_corr.StatisticalFirstOrder.transform_sigma_points
+        )
         sigma_points, _, _ = sigma_points_fn(self._mm, extrapolated.hidden_state)
 
         fx = vmap_f(sigma_points.T).T  # (d, S).T = (S, d) -> (S, d) -> transpose again
-        center_fn = jax.vmap(_scalar.StatisticalFirstOrder.center)
+        center_fn = jax.vmap(scalar_corr.StatisticalFirstOrder.center)
         fx_mean, _, fx_centered_normed = center_fn(self._mm, fx)
 
         # Compute output scale and error estimate
-        calibrate_fn = jax.vmap(_scalar.StatisticalFirstOrder.calibrate)
+        calibrate_fn = jax.vmap(scalar_corr.StatisticalFirstOrder.calibrate)
         error_estimate, output_scale_sqrtm = calibrate_fn(
             self._mm, fx_mean, fx_centered_normed, extrapolated.hidden_state
         )
@@ -86,23 +89,27 @@ class BlockDiagStatisticalFirstOrder(_collections.AbstractCorrection):
 
         H, noise = self.linearize(extrapolated, vmap_f)
 
-        fn = jax.vmap(_scalar.StatisticalFirstOrder.complete_correction_post_linearize)
+        fn = jax.vmap(
+            scalar_corr.StatisticalFirstOrder.complete_correction_post_linearize
+        )
         return fn(self._mm, H, extrapolated.hidden_state, noise)
 
     def linearize(self, extrapolated, vmap_f):
         # Transform the sigma-points
-        sigma_points_fn = jax.vmap(_scalar.StatisticalFirstOrder.transform_sigma_points)
+        sigma_points_fn = jax.vmap(
+            scalar_corr.StatisticalFirstOrder.transform_sigma_points
+        )
         sigma_points, _, sigma_points_centered_normed = sigma_points_fn(
             self._mm, extrapolated.hidden_state
         )
 
         # Evaluate the vector field at the sigma-points
         fx = vmap_f(sigma_points.T).T  # (d, S).T = (S, d) -> (S, d) -> transpose again
-        center_fn = jax.vmap(_scalar.StatisticalFirstOrder.center)
+        center_fn = jax.vmap(scalar_corr.StatisticalFirstOrder.center)
         fx_mean, _, fx_centered_normed = center_fn(self._mm, fx)
 
         # Complete the linearization
-        lin_fn = jax.vmap(_scalar.StatisticalFirstOrder.linearization_matrices)
+        lin_fn = jax.vmap(scalar_corr.StatisticalFirstOrder.linearization_matrices)
         return lin_fn(
             self._mm,
             fx_centered_normed,
@@ -119,25 +126,25 @@ _TS0CacheType = Tuple[jax.Array]
 class BlockDiagTaylorZerothOrder(_collections.AbstractCorrection):
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
-        self._ts0 = _scalar.TaylorZerothOrder(*args, **kwargs)
+        self._ts0 = scalar_corr.TaylorZerothOrder(*args, **kwargs)
 
     def begin_correction(self, x, /, vector_field, t, p):
-        select_fn = jax.vmap(_scalar.TaylorZerothOrder.select_derivatives)
+        select_fn = jax.vmap(scalar_corr.TaylorZerothOrder.select_derivatives)
         m0, m1 = select_fn(self._ts0, x.hidden_state)
 
         fx = vector_field(*m0.T, t=t, p=p)
 
-        marginalise_fn = jax.vmap(_scalar.TaylorZerothOrder.marginalise_observation)
+        marginalise_fn = jax.vmap(scalar_corr.TaylorZerothOrder.marginalise_observation)
         cache, obs_unbatch = marginalise_fn(self._ts0, fx, m1, x.hidden_state)
 
-        mahalanobis_fn = _scalar.NormalQOI.mahalanobis_norm
+        mahalanobis_fn = scalar_vars.NormalQOI.mahalanobis_norm
         mahalanobis_fn_vmap = jax.vmap(mahalanobis_fn)
         output_scale_sqrtm = mahalanobis_fn_vmap(obs_unbatch, jnp.zeros_like(m1))
         error_estimate = obs_unbatch.cov_sqrtm_lower
         return output_scale_sqrtm * error_estimate, output_scale_sqrtm, cache
 
     def complete_correction(self, extrapolated, cache: _TS0CacheType):
-        fn = jax.vmap(_scalar.TaylorZerothOrder.complete_correction)
+        fn = jax.vmap(scalar_corr.TaylorZerothOrder.complete_correction)
         return fn(self._ts0, extrapolated, cache)
 
     def _cov_sqrtm_lower(self, cov_sqrtm_lower):
